refactor: move state to a dir
